ConcertReader의 register 기능 -> Writer 클래스로 분리
